+++ backend/src/app/rpc/commands/files_update.clj
@@ -93,6 +93,15 @@
    [:features {:optional true} ::cfeat/features]])
 
 (def ^:private
+  schema:rename-file-page
+  [:map {:title "rename-file-page"}
+   [:id ::sm/uuid]
+   [:page-id ::sm/uuid]
+   [:name [:string {:max 250}]]
+   [:session-id {:optional true} ::sm/uuid]
+   [:features {:optional true} ::cfeat/features]])
+
+(def ^:private
   schema:page-summary
   [:map {:title "PageSummary"}
    [:id ::sm/uuid]
@@ -101,6 +110,14 @@
 (def ^:private
   schema:create-file-page-result
   [:map {:title "create-file-page-result"}
+   [:file-id ::sm/uuid]
+   [:page schema:page-summary]
+   [:revn {:min 0} ::sm/int]
+   [:vern {:min 0} ::sm/int]])
+
+(def ^:private
+  schema:rename-file-page-result
+  [:map {:title "rename-file-page-result"}
    [:file-id ::sm/uuid]
    [:page schema:page-summary]
    [:revn {:min 0} ::sm/int]
@@ -355,6 +372,60 @@
                          (cfeat/check-client-features! (:features params))
                          (cfeat/check-file-features! (:features file)))
         page-request (headless/create-page-request (blob/decode (:data file)) params)
+        changes      (:changes page-request)
+        session-id   (or session-id (uuid/next))
+        cfg          (assoc cfg ::timestamp (ct/now))
+        update-args  {:id id
+                      :revn (:revn file)
+                      :vern (:vern file)
+                      :file file
+                      :team team
+                      :features (set/difference features cfeat/frontend-only-features)
+                      :changes changes
+                      :session-id session-id
+                      :profile-id profile-id}]
+
+    (mtx/run! metrics {:id :update-file-changes :inc (count changes)})
+    (update-file* cfg update-args)
+
+    (with-meta {:file-id id
+                :page (:page page-request)
+                :revn (inc (:revn file))
+                :vern (:vern file)}
+      {::audit/replace-props
+       {:id         (:id file)
+        :name       (:name file)
+        :features   (:features file)
+        :project-id (:project-id file)
+        :team-id    (:team-id file)}})))
+
+(sv/defmethod ::rename-file-page
+  {::climit/id [[:update-file/by-profile ::rpc/profile-id]
+                [:update-file/global]]
+
+   ::webhooks/event? true
+   ::webhooks/batch-timeout (ct/duration "2m")
+   ::webhooks/batch-key (webhooks/key-fn ::rpc/profile-id :id)
+
+   ::sm/params schema:rename-file-page
+   ::sm/result schema:rename-file-page-result
+   ::doc/module :files
+   ::doc/added "2.15.4"
+   ::db/transaction true}
+  [{:keys [::mtx/metrics ::db/conn] :as cfg}
+   {:keys [::rpc/profile-id id session-id] :as params}]
+
+  (files/check-edition-permissions! conn profile-id id)
+  (db/xact-lock! conn id)
+
+  (let [file         (get-file cfg id)
+        team         (teams/get-team conn
+                                     :profile-id profile-id
+                                     :team-id (:team-id file))
+        features     (-> (cfeat/get-team-enabled-features cf/flags team)
+                         (cfeat/check-client-features! (:features params))
+                         (cfeat/check-file-features! (:features file)))
+        page-request (headless/rename-page-request (blob/decode (:data file)) params)
         changes      (:changes page-request)
         session-id   (or session-id (uuid/next))
         cfg          (assoc cfg ::timestamp (ct/now))